fix: Multiple renders of confirm delete modal
When trying to delete an item. Modals of all other items show because they
all rely on a single piece of information in the redux store. They all toggle
when on of them is being changed state. In this commit, this behavior is
prevented by having additional condition on when to show the modal. Delete
confirm modals now only show one at a time.