feat: Post item detail page added
Add PostItemDetailPage. The users are now able to expand details of a certain
post by clicking on the list item. Routes to specific posts are also available.
As of now, there are no interactions when users land to the detail page. But
in the next commits, other features such as commenting and liking will be
added.