feat: Display post comments in post detail
Retrieve comments from the server and display them when a post is opened
and the user is redirected to the opened post link. The comments can be
upvoted and downvoted just like the posts. This is made possible by refactoring
VotingCircle component to accept callbacks for its buttons as the props.