feat: Activate toggle mechanism for showing comment form
Users can now toggle between showing and hiding the comment form. The
comment form is hidden in default so the user will have to click on the reply
button provided on the post. When the form shows up, the user can click the
reply button again to hide the form or hit cancel in the comment form.