fix: Keep sort order after submitting a post
The posts list are not in the selected sort order when submitting post because
the new posts always gets to the bottom of the list no matter what the sort
order is. The way this is fixed is to call sortPosts whenever the user adds a post
so that the list will be in the right order. To make components aware of the
changes of the list in the redux store, the reducer for sorting posts is also modified
such that it does not modify the existing copy of the list but instead create
a new sorted copy.