refactor: Combine action creators and thunks
This is done to make the code cleaner and more concise. Action creators
tend to grow in number and they basically have the same intention as the
thunks. In this commit, instead of dispatching action creators in thunks, the
action objects themselves are dispatched. This might have some side effects
like losing action creators' reusability.