fix: Errors in toggles
Toggles for forms for editing and deleting posts and comments throw errors.
The reducers for these actions is not updated when editingPost and editingComment
were deleted from the redux store. Thus, these actions rely on undefined
values that cause errors in performing the actions. In this commit, these are
fixed.