style: Disable some input fields when editing post
Some of post details can not be edited by the user once they are posted. So,
in this commit, the input fields for those properties are disabled since they
are not allowed to be edited in the server anyways. This will give the users
hint about the restrictions when editing a post.